// web/src/shared/api/generated/model/getPaymentLogParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PaymentStatus } from './paymentStatus';
import type { SubscriptionPlan } from './subscriptionPlan';

export type GetPaymentLogParams = {
status?: PaymentStatus;
plan?: SubscriptionPlan;
studentId?: string;
reference?: string;
from?: string;
to?: string;
needsReview?: boolean;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
pageNumber?: number | string;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
pageSize?: number | string;
};